fix: Restructure registry (#180)
