+++ app/src/config/index.tsx
@@ -8,6 +8,7 @@ import { mainnet, sepolia } from 'wagmi/chains'
 if (!projectId) throw new Error('Project ID is not defined')
 
 const vercelUrl = vercelDomain ? `https://${vercelDomain}` : ''
+export const DWELLIR_KEY = process.env.NEXT_PUBLIC_DWELLIR_KEY
 
 const metadata = {
   name: 'turtle-app',
