+++ app/src/utils/paraspell.ts
@@ -1,4 +1,4 @@
-import { getAssetUid, isAssetHub } from '@/config/registry'
+import { getAssetUid, isAssetHub } from '@/registry'
 import { TransferParams } from '@/hooks/useTransfer'
 import { Chain } from '@/models/chain'
 import { Token } from '@/models/token'
